pyproject.toml
[project]
name = "autogen"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "autogen-agentchat",
    "autogen-ext",
    "autogen-ext[openai,azure]",
    "mcp>=1.10.0",
    "python-dotenv",
    "tavily-python",
    "aiohttp",
]
